digifinex: convert to TS
